# src/backend/openapi.yaml
openapi: 3.0.3
info:
  title: LearnFlow API
  version: 0.1.0
  description: >
    Verbindliche Schnittstellen-Definition (API-First, ADR-010). Grundgerüst (T-03):
    Endpunkte aus US-01..US-05 als Stubs. Auth (US-05) und Dokument-Upload (US-04) sind
    mit vollständigen Request/Response-Schemas ausmodelliert; die übrigen Endpunkte sind
    bewusst leere Stubs und werden in den jeweiligen Folge-Issues ausmodelliert.
servers:
  - url: /
...
security:
  - bearerAuth: []
paths:
...
  /api/documents:
    get:
      tags: [documents]
      summary: Dokumente des eigenen Bereichs auflisten (US-04)
      responses:
        '200':
          description: Liste der Dokumente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentList'
        '401':
          description: Nicht authentifiziert
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Keine Berechtigung (Rolle knowledge_owner erforderlich)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags: [documents]
      summary: Dokument hochladen; gleicher Dateiname ersetzt das bestehende (US-04, T-15)
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/DocumentUploadRequest'
      responses:
        '200':
          description: >-
            Bestehendes Dokument gleichen Namens ersetzt — gleiche id, alte Chunks
            gelöscht, Neu-Indexierung läuft
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentResponse'
        '201':
          description: Neues Dokument angelegt, Verarbeitung läuft
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentResponse'
        '401':
          description: Nicht authentifiziert
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Keine Berechtigung (Rolle knowledge_owner erforderlich)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '413':
          description: Datei überschreitet das 10-MB-Limit
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '415':
          description: 'Nicht unterstütztes Dateiformat (erlaubt: PDF, DOCX, Markdown)'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '400':
          description: 'Unbekannter Bereich (MVP: nur der hartcodierte Pilot-Bereich)'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/documents/{document_id}:
    get:
      tags: [documents]
      summary: Dokument abrufen inkl. Verarbeitungsstatus (US-04)
      parameters:
        - name: document_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Dokument mit aktuellem Status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentResponse'
        '401':
          description: Nicht authentifiziert
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Keine Berechtigung (Rolle knowledge_owner erforderlich)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Dokument nicht gefunden (auch bei fremdem Bereich)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      tags: [documents]
      summary: Dokument löschen (US-04)
      parameters:
        - name: document_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Dokument gelöscht
        '401':
          description: Nicht authentifiziert
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Keine Berechtigung (Rolle knowledge_owner erforderlich)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Dokument nicht gefunden (auch bei fremdem Bereich)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/documents/{document_id}/content:
    get:
      tags: [documents]
      summary: Kontextfenster um einen zitierten Chunk abrufen, für den Quellenlink-Viewer (US-01, T-21)
      description: >
        Für jede Rolle erreichbar, nicht nur `knowledge_owner`: eine Antwort auf
        `/api/query` zitiert bereits den Inhalt des Chunks unter `around`, ein
        Lernender darf also öffnen, was ihm ohnehin zitiert wurde. Deshalb ist
        `around` Pflicht und die Antwort ist auf ein Fenster um diesen Chunk
        begrenzt — anders als der zitierte Chunk hat der Rest des Dokuments dem
        Retrieval-Gate für diese Frage nie genügt, und der Endpoint liefert ihn
        deshalb nicht mit. Liefert den geparsten Text, nicht die hochgeladene
        Originaldatei (ADR-003) — der belegende Abschnitt ist der Chunk, und nur
        damit ist "hervorgehoben" ohne Koordinaten-Rückrechnung exakt umsetzbar.
      parameters:
        - name: document_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: around
          in: query
          required: true
          description: chunk_id einer Citation (US-01) — der Chunk, um den das Fenster liegt.
          schema:
            type: string
            format: uuid
        - name: window
          in: query
          required: false
          description: >
            Chunks vor und nach `around`, die zusätzlich mitgeliefert werden
            (US-01 "Abschnitt" — Kontext für den belegenden Chunk, nicht das
            ganze Dokument).
          schema:
            type: integer
            minimum: 0
            maximum: 10
            default: 2
      responses:
        '200':
          description: Fenster von Chunks um `around`, in Lesereihenfolge
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentContent'
        '401':
          description: Nicht authentifiziert
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: >
            Dokument nicht gefunden (auch bei fremdem Bereich), oder `around`
            gehört nicht zu diesem Dokument.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: >
            Dokument ist noch nicht `available` (Status `pending`, `processing`
            oder `failed`) — es gibt noch keine oder keine gültigen Chunks zu
            zeigen.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Error:
      type: object
      required: [detail]
      properties:
        detail:
          type: string
          description: Menschenlesbare Fehlerbeschreibung
...
    DocumentStatus:
      type: string
      enum: [pending, processing, available, failed]
    DocumentUploadRequest:
      type: object
      required: [file]
      properties:
        file:
          type: string
          format: binary
          description: Dokument (PDF, DOCX oder Markdown), max. 10 MB
        area:
          type: string
          description: Bereich; im MVP hartcodierter Pilot-Bereich
          default: default
    DocumentResponse:
      type: object
      required: [id, filename, status, area, chunk_count, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        filename:
          type: string
        status:
          $ref: '#/components/schemas/DocumentStatus'
        area:
          type: string
        chunk_count:
          type: integer
        error_message:
          type: string
          nullable: true
        created_at:
          description: Zeitpunkt des ersten Uploads; überlebt eine Ersetzung (T-15)
          type: string
          format: date-time
        updated_at:
          description: Zeitpunkt des Uploads der aktuellen Fassung (T-15)
          type: string
          format: date-time
        uploaded_by:
          description: >-
            E-Mail des Hochladenden der aktuellen Fassung (#92); null, wenn das
            Konto seither gelöscht wurde (FK ON DELETE SET NULL). Erlaubt der
            Upload-UI, vor einer Ersetzung den bisherigen Hochladenden zu nennen.
          type: string
          nullable: true
    DocumentList:
      type: array
      items:
        $ref: '#/components/schemas/DocumentResponse'
    DocumentContentChunk:
      type: object
      description: Ein Chunk in Lesereihenfolge (T-21)
      required: [chunk_id, chunk_index, page, heading, content]
      properties:
        chunk_id:
          type: string
          format: uuid
        chunk_index:
          type: integer
        page:
          type: integer
          nullable: true
        heading:
          type: string
          nullable: true
        content:
          type: string
    DocumentContent:
      type: object
      description: Geparster Dokumentinhalt für den Quellenlink-Viewer (US-01, T-21)
      required: [id, filename, status, chunks]
      properties:
        id:
          type: string
          format: uuid
        filename:
          type: string
        status:
          $ref: '#/components/schemas/DocumentStatus'
        chunks:
          type: array
          items:
            $ref: '#/components/schemas/DocumentContentChunk'
          description: Sortiert nach `chunk_index`
...